[mod] improve cache performanche
Now it is possible to set a different cache limit for each single service.
If set it overwrites the global cache limit configuration. Second, the
cache invalidation time is randomly shifted by a displacement factor. This
means that if you have a cache limit of 20 min and a displacement factor of
o.5, the invalidation time is 10 to 30 min. Now, the moment of necessary
service data updates is homogenously distributed and not all service updates
will be fetched together in one single moment, how it has been until now.